1 more

+++ include/functions.h
@@ -205,7 +205,7 @@ u8 func_80027E84(u8 arg0, u8 arg1, f32 arg2);
 // ? func_800281E8(?);
 // ? func_8002829C(?);
 // ? func_80028304(?);
-void func_8002836C(GlobalContext*, Vec3f*, Vec3f*, Vec3f*, Color_RGBA8*, Color_RGBA8*, s32, s32, s32);
+void func_8002836C(GlobalContext*, Vec3f*, Vec3f*, Vec3f*, Color_RGBA8*, Color_RGBA8*, s16, s32, s32);
 // ? func_800283D4(?);
 // ? func_8002843C(?);
 // ? func_800284A4(?);
